deployer233332_render_requirements.txt
aiohttp>=3.12.15
python-dotenv>=1.1.1
pyyaml>=6.0.2
telethon>=1.40.0